+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## XX.XX.XX
+* Added support for Remote Config feature accesible through "Countly.Instance.RemoteConfig()" interface:
+  * "DownloadKeys" for fetching RC values from server
+  * "GetValues" for accessing all RC values
+  * "GetValue" for accessing the given RC value
+* Added configuration option "EnableRemoteConfigAutomaticTriggers" to automatically download remote config values:
+  * After initialization finishes.
+  * RemoteConfig consent is given.
+  * After the device ID is changed to a different user.
+
 ## 25.4.3
 * ! Minor breaking change ! User properties will now be automatically saved under the following conditions:
   * When an event is recorded